improved pack feature
