+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "leetcode-runner-cli"
-version = "0.2.2"
+version = "0.2.5"
 edition = "2021"
 
 [dependencies]
