+++ src/main.rs
@@ -1,6 +1,6 @@
 use crate::args::Cli;
 use crate::file_parser::codefile::CodeFile;
-use crate::utils::{execute_testcases, submit};
+use crate::utils::{execute_testcases, pack, submit};
 
 use args::Commands;
 use clap::Parser;
@@ -90,47 +90,8 @@ fn main() -> Result<()> {
                 bail!("Aborting submission due to failed testcase(s)".red().bold());
             }
         }
-        Some(Commands::Pack { file }) => {
-            let code_file = if let Some(path) = file {
-                CodeFile::from_file(&path)?
-            } else {
-                CodeFile::from_dir(".")?
-            };
-            let question = lc.question_content(&code_file.question_title)?;
+        Some(Commands::Pack { file }) => pack(&lc, file)?,
 
-            // create a directory if it doesn't exists with name of question
-            // create a README.md file with the question description
-            // create a file with the code
-            std::fs::create_dir_all(&code_file.question_title.replace(' ', ""))?;
-
-            std::fs::write(
-                format!(
-                    "{}/main.{}",
-                    &code_file.question_title,
-                    code_file.language.extension()
-                ),
-                code_file.code,
-            )?;
-
-            // dont create readme if it exists
-            if let Ok(mut readme_file) = std::fs::OpenOptions::new()
-                .create_new(true)
-                .write(true)
-                .open(format!("{}/{}", &code_file.question_title, GIT_README))
-            {
-                println!(
-                    "You can write your notes about question in {}/README.md",
-                    &code_file.question_title
-                );
-                std::io::Write::write_all(
-                    &mut readme_file,
-                    format!("# {}\n", code_file.question_title).as_bytes(),
-                )?;
-                std::io::Write::write_all(&mut readme_file, question.content.as_bytes())?;
-            } else {
-                println!("{} already exists, skipping creation.", GIT_README);
-            };
-        }
         None => {}
     };
